+++ app/Http/Controllers/Api/SoldHouseController.php
@@ -8,44 +8,62 @@ use App\Http\Resources\SoldHouseResource;
 use App\Models\Apartment;
 use App\Models\SoldHouse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
 
 class SoldHouseController extends Controller
 {
     public function index()
     {
-        return SoldHouseResource::collection(SoldHouse::get());
+        if (auth()->user()->hasAnyRoles(['super-admin', 'accountant', 'manager'])) {
+            return SoldHouseResource::collection(SoldHouse::get());
+        }
+        $this->messageNotAllowedTo();
     }
 
     public function getById($id)
     {
-        return new SoldHouseResource(SoldHouse::find($id));
+        if (auth()->user()->hasAnyRoles(['super-admin', 'accountant', 'manager'])) {
+            return new SoldHouseResource(SoldHouse::find($id));
+        }
+        $this->messageNotAllowedTo();
     }
 
     public function create(SoldHouseRequest $request)
     {
-        $data = $request->validated();
-        $square = Apartment::query()
-            ->where('id', '=', $data['apartment_id'])
-            ->value('square');
-        $price = Apartment::query()
-            ->where('id', '=', $data['apartment_id'])
-            ->value('price');
+        if (auth()->user()->hasRole('super-admin')){
+            $data = $request->validated();
+            $square = Apartment::query()
+                ->where('id', '=', $data['apartment_id'])
+                ->value('square');
+            $price = Apartment::query()
+                ->where('id', '=', $data['apartment_id'])
+                ->value('price');
 
-        $data['summa'] = $square * $price;
+            $data['summa'] = $square * $price;
 
-        SoldHouse ::create($data);
-        return $this->createdData();
+            return $this->createdData([
+                'soldHouse' => SoldHouse ::create($data)
+            ]);
+        }
+        $this->messageNotAllowedTo();
     }
 
     public function update(SoldHouseRequest $request, SoldHouse $soldHouse)
     {
-        $soldHouse->update($request->all());
-        return $this->updatedData();
+        if (auth()->user()->hasRole('super-admin')){
+            return $this->updatedData([
+                'soldHouse' => $soldHouse->update($request->validated())
+            ]);
+        }
+        $this->messageNotAllowedTo();
     }
 
     public function destroy(SoldHouse $soldHouse)
     {
-        $soldHouse->delete();
-        return $this->deletedData();
+        if (auth()->user()->hasRole('super-admin')){
+            $soldHouse->delete();
+            return $this->deletedData();
+        }
+        $this->messageNotAllowedTo();
     }
 }
